# config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "JavaInUse API reference for developers",
    "version": "1.0",
    "title": "JavaInUse API",
    "termsOfService": "http://javainuse.com",
    "contact": {
      "name": "Mokhtar MEHALI",
      "url": "http://asp-public.fr",
      "email": "mokhtar.mehali@asp-public.fr"
    },
    "license": {
      "name": "JavaInUse License",
      "url": "javainuse@gmail.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
    "/api/hello": {
      "get": {
        "tags": [
          "hello-controller"
        ],
        "summary": "sayHello",
        "operationId": "sayHelloUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/rest/hello": {
      "get": {
        "tags": [
          "hello-resource"
        ],
        "summary": "Returns Hello World",
        "operationId": "helloUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "100": {
            "description": "100 is the message"
          },
          "200": {
            "description": "Successful Hello World",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
...
}